+++ src/main/java/tech/allegro/wildsnake/showcase/model/ShowcaseItem.java
@@ -1,11 +1,13 @@
 package tech.allegro.wildsnake.showcase.model;
 
+import java.math.BigDecimal;
+
 public class ShowcaseItem {
     private final String title;
     private final String imageUrl;
-    private final int price;
+    private final BigDecimal price;
 
-    public ShowcaseItem(String title, String imageUrl, int price) {
+    public ShowcaseItem(String title, String imageUrl, BigDecimal price) {
         this.title = title;
         this.imageUrl = imageUrl;
         this.price = price;
@@ -19,7 +21,7 @@ public class ShowcaseItem {
         return imageUrl;
     }
 
-    public int getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 }
